src/requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.5.0.5
cachetools==4.1.0
celery==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
discovery==0.0.4
Django==3.0.5
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.5.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0